Initial Submit - Version 1.0.0.1
This is the initial submit for the TSTP Mail Auth Tester.

This program is designed to allow you to test email authentication with Microsoft, Google, Yahoo, or a custom server, to see if your e-mail credentials and settings are valid for use.  This program will check your credentials and send a test email if successful to the recipient added or to TSTP if no recipient is provided.  If the authentication test fails, it gives you a message that shows you why it failed so that you can look into the issue to fix it.

This program has an extensive tutorial that shows you the settings to put, security settings that may need to be implemented or disabled depending on your server, and some helpful information to help give you a better chance at a successful test.

The full, compiled, version of this program can be downloaded at the TSTP.xyz website.

Edit and make this your own for your specific needs.